% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ccr_setup}
\alias{ccr_setup}
\title{Call this function when it's your first time installing this package.
This will set up the miniconda environment for python,
which is how this R package works under the hood.}
\usage{
ccr_setup()
}
\value{
No return value, called for side effects.
}
\description{
Call this function when it's your first time installing this package.
This will set up the miniconda environment for python,
which is how this R package works under the hood.
}
\examples{
ccr_setup()

}
