% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccrFit.R
\name{ccrFit}
\alias{ccrFit}
\title{Fit the correspondence curve regression}
\usage{
ccrFit(formula, data, model = TRUE, link = c("loglog", "logitlogit"),
  is.slope = FALSE, is.ht = FALSE, par.ini, tm = seq(0.01, 0.9999,
  length.out = 50), method = "ccr.fit", sig.level = 0.05)
}
\arguments{
\item{formula}{object of class 'formula'
(or one that can be coerced to that class):
a symbolic description of the model to be fitted.
The details of model specification are given under 'Details'.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)
containing the variables in the model. If not found in data,
the variables are taken from environment(formula),
typically the environment from which ccrFit is called.}

\item{model}{logical. If TRUE the corresponding components of the fit  are returned.}

\item{link}{The link function to be used for fitting, currently only 
link = 'loglog' and 'logitlogit'.}

\item{is.slope}{logical. Only for link 'logitlogit'. Default is FALSE.}

\item{is.ht}{logical. Only for link 'logitlogit'. Default is FALSE.}

\item{par.ini}{the initial values for the estimate parameters.
If is.null(par.ini) == TRUE, par.ini is set in the the details.}

\item{tm}{The vector of tm.}

\item{method}{The method to be used for fitting, currently only method = 'ccr.fit'.}

\item{sig.level}{The significant level. Default is 0.05.}
}
\value{
A list with the elements:

\describe{
  \item{coefficients}{a named vector of coefficients.}
  \item{model}{if requested (the default), the model frame used.}
  \item{call}{the matched call.}
  \item{std.error}{the estimated standard errors.}
  \item{CI}{the confidence intervals.}
  }
}
\description{
Fit the correspondence curve regression
}
\details{
Please refer to \strong{Li, Q. and Zhang, F. (2017)}.
}
\examples{
## The example of ChIP data
\dontrun{
data(data_ChIP)
## estimate
par.ini = c(1, 0, -0.1, 0, 0, 0.5)  # initial value
fit <- ccrFit(cbind(y1, y2)~x, data = data_ChIP,
             link = 'loglog', par.ini = par.ini)
             
             
## The example of microarray data
data(data_microarray)
## with slope without ht
par.ini = c(-1, 1, 0.1, 0.1)    # initial value
fit <- ccrFit(cbind(y1,y2)~x1+x2, 
              data = data_microarray,
              link = 'logitlogit', 
              is.slope = TRUE, is.ht = FALSE, 
              par.ini = par.ini)

## with slope with ht, including iteractions
par.ini = c(-1, 0.1, 0.1, 0, 1,  0.1, 0.1, 0)
fit <- ccrFit(cbind(y1,y2)~x1*x2, 
              data = data_microarray,
              link = 'logitlogit', 
              is.slope = TRUE, is.ht = TRUE, 
              par.ini = par.ini)
}
}
\references{
Li, Q. and Zhang, F. (2017). 
A regression framework for assessing covariate effects 
on the reproducibility of high-throughput experiments. 
Biometrics, \url{http://dx.doi.org/10.1111/biom.12832}.
}
\author{
Feipeng Zhang and Qunhua Li
}
\keyword{ccrFit}
