% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr_run_shiny.R
\name{ccr_shiny}
\alias{ccr_shiny}
\title{Launches a user-friendly interface (shiny app) for this package}
\usage{
ccr_shiny()
}
\value{
No return value, called for side effects.
}
\description{
Launches a user-friendly interface (shiny app) for this package
}
