% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\name{ccr_wrapper}
\alias{ccr_wrapper}
\title{CCR wrapper function}
\usage{
ccr_wrapper(data_file, data_col, q_file, q_col, model = "all-MiniLM-L6-v2")
}
\arguments{
\item{data_file}{Name of the csv file of user-supplied data, or an R data frame}

\item{data_col}{Name of the relevant data column}

\item{q_file}{Name of the csv file of questionnaire data, or an R data frame}

\item{q_col}{Name of the questionnaire column}

\item{model}{Name of a huggingface model (https://huggingface.co/models); "all-MiniLM-L6-v2" by default}
}
\value{
A data frame with similarity score columns appended
}
\description{
CCR wrapper function
}
\examples{
ccr_wrapper("data/test.csv", "d", "data/test.csv", "q")
}
